fix electron custom url handling
