+++ electron/src/trusted-main-electron-process/CommandLineApi.ts
@@ -12,23 +12,6 @@ export const writeResultToStdOutAndExit = (result: string) => {
   process.exit(0)  
 }
 
-export const getCommandLineArguments = (): string[] => {
-  /**
-    When executed with electron command the first two arguments are the path and the executable name,
-    so the actual command line arguments start at index 2. When executed as packed binary the actual
-    command line arguments start at index 1.
-    See: https://nodejs.org/docs/latest/api/process.html#process_process_argv.
-    The constant represents only the arguments that follow the path and executable name.
-  */
-  const argv = process.argv
-  if(argv[0].toLowerCase().endsWith("electron")){
-      return argv.slice(2)
-  }else{
-      return argv.slice(1)
-  }
-};
-
-
 // IpcApiFactory.implementSyncApi( "getCommandLineArguments", () => {
 //   /**
 //     When executed with electron command the first two arguments are the path and the executable name,
